Add files via upload
This project uses the tkinter module in Python to create a basic interactive chatbot application. The Chatbot converses with users, responding to their questions about return policies, technical help, and product information. Regular expressions are used by the program to match user inputs with specified intents and produce relevant responses.

Take note:

Please be advised that this application's graphical user interface (GUI) is not intended for aesthetics; rather, it is intended primarily for demonstration. It is a useful illustration of how a chatbot communicates with users. This project's main goal is to demonstrate the Chatbot system's logic and operation, not its aesthetic appeal.