Add files via upload
A Python-based program driven by natural language processing and machine learning is called the Movie Recommendation System. This project provides an interactive interface that allows users to find comparable movies depending on their tastes by utilizing the Streamlit framework. The system works with a movie dataset, pulling out attributes like cast and crew as well as genres. When a user chooses a certain movie, the application uses cosine similarity to calculate how similar the two films are, then suggests the top five titles that are most similar. The user experience is improved by having the recommendations shown next to the corresponding posters. The idea offers movie buffs an easy and effective approach to discover new films that fit their preferences.