TicTacToe
This project uses the minimax algorithm to implement a classic game of Tic Tac Toe against an AI opponent. The p5.js library was used to develop the graphical user interface (GUI) for the game. The AI responds to actions made by players by clicking on the grid, ensuring that the gaming is challenging. The motions are based on the minimax algorithm. When a player wins or there is a tie, the application finishes the game and visually displays the winning combinations. In addition to offering players a fun experience, this strategic and engaging game highlights the potential of artificial intelligence in gaming environments.